src: add lang::DelimitedLanguageRegister
This class is not only a Dictionary, it can check a whole text at once.
It's based on a dict::DynamicDictionary with lang::LanguageSeparator addition
A separator can be declare with prefix/suffix/no space surrounding.

This class is not done implementing yes, it lacks at least an isExistingEntry
overriding.

Added tests accordingly